requirements.txt
alembic==0.8.8
aniso8601==1.1.0
click==6.6
coverage==4.2
coveralls==1.1
docopt==0.6.2
factory-boy==2.7.0
fake-factory==0.7.2
Flask==0.11.1
Flask-Login==0.3.2
Flask-Migrate==2.0.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
mysqlclient==1.3.9
passlib==1.6.5
pluggy==0.3.1
py==1.4.31
pycrypto==2.6.1
PyJWT==1.4.2
python-dateutil==2.5.3
python-editor==1.0.1
pytz==2016.6.1
requests==2.11.1
six==1.10.0
SQLAlchemy==1.0.15
SQLAlchemy-Utils==0.32.9
tox==2.3.1
virtualenv==15.0.3
Werkzeug==0.11.11
